Move SingleTrigger to the TriggersRow folder
