+++ src/features/CourtCaseList/CourtCaseListEntry/TriggersRow/SingleTrigger.tsx
