+++ src/features/CourtCaseList/CourtCaseListEntry/TriggersRow/TriggersRow.tsx
@@ -3,7 +3,7 @@ import { Table } from "govuk-react"
 import { createUseStyles } from "react-jss"
 import { gdsLightGrey } from "utils/colours"
 import Image from "next/image"
-import { SingleTrigger } from "../SingleTrigger"
+import { SingleTrigger } from "./SingleTrigger"
 import LockedByTag from "features/CourtCaseList/tags/LockedByTag/LockedByTag"
 import CaseUnlockedTag from "features/CourtCaseList/tags/CaseUnlockedTag"
 import Trigger from "services/entities/Trigger"
